refactor(teams): module 2 - extract panel components, fix FormData type cast
