+++ src/app/(app)/team-management.tsx
@@ -2,9 +2,9 @@ import Feather from '@expo/vector-icons/Feather';
 import * as ImagePicker from 'expo-image-picker';
 import { useRouter } from 'expo-router';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import { Alert, Pressable, Share, TextInput, View } from 'react-native';
+import { Alert, Pressable, View } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
-import { Button, Card, Spinner } from 'heroui-native';
+import { Card } from 'heroui-native';
 
 import { AppText } from '../../components/app-text';
 import { ScreenScrollView } from '../../components/screen-scroll-view';
@@ -13,11 +13,17 @@ import { SectionLabel } from '../../components/section-label';
 import { mflColors } from '../../constants/colors';
 import { useLeagueContext } from '../../contexts/league-context';
 import { useRole } from '../../contexts/role-context';
-import { ActionButton } from '../../features/team-management/components/action-button';
 import { ManagedTeamCard } from '../../features/team-management/components/managed-team-card';
-import { MemberListRow } from '../../features/team-management/components/member-list-row';
 import { TeamManagementHeader } from '../../features/team-management/components/team-management-header';
+import { AddMembersPanel } from '../../features/team-management/components/add-members-panel';
 import { TeamManagementPanel } from '../../features/team-management/components/team-management-panel';
+import { CaptainPanel } from '../../features/team-management/components/captain-panel';
+import { GovernorsPanel } from '../../features/team-management/components/governors-panel';
+import { InvitePanel } from '../../features/team-management/components/invite-panel';
+import { MembersPanel } from '../../features/team-management/components/members-panel';
+import { TeamNameForm } from '../../features/team-management/components/team-name-form';
+import { UnallocatedPanel } from '../../features/team-management/components/unallocated-panel';
+import { ViceCaptainPanel } from '../../features/team-management/components/vice-captain-panel';
 import {
   useManagedTeamMembers,
   useTeamManagementActions,
@@ -28,12 +34,7 @@ import type {
   ManagedTeam,
   ManagedTeamMember,
   PickedTeamLogo,
-  TeamManagementData,
 } from '../../features/team-management/types/team-management';
-import {
-  getPerTeamCapacity,
-  searchMembers,
-} from '../../features/team-management/utils/team-management-utils';
 
 type Panel =
   | 'create'
@@ -47,16 +48,6 @@ type Panel =
   | 'invite'
   | null;
 
-const inputStyle = {
-  backgroundColor: mflColors.card,
-  borderWidth: 1,
-  borderColor: mflColors.border,
-  borderRadius: 12,
-  paddingHorizontal: 16,
-  paddingVertical: 12,
-  fontSize: 16,
-  color: mflColors.text,
-};
 
 export default function TeamManagementScreen() {
   const router = useRouter();
@@ -153,11 +144,13 @@ export default function TeamManagementScreen() {
 
     if (result.canceled || !result.assets?.[0]) return;
     const asset = result.assets[0];
-    const rawType = (asset as { mimeType?: string }).mimeType ?? asset.type;
+    const assetWithMime = asset as typeof asset & { mimeType?: string };
     const file: PickedTeamLogo = {
       uri: asset.uri,
       name: asset.fileName ?? `team_logo_${Date.now()}.jpg`,
-      type: rawType?.startsWith('image/') ? rawType : 'image/jpeg',
+      type: assetWithMime.mimeType?.startsWith('image/')
+        ? assetWithMime.mimeType
+        : 'image/jpeg',
     };
 
     actions.uploadLogo.mutate(
@@ -680,606 +673,4 @@ export default function TeamManagementScreen() {
       );
     }
   }
-}
-
-function TeamNameForm({
-  value,
-  onChange,
-  onSubmit,
-  buttonLabel,
-  isBusy,
-  disabled = false,
-}: {
-  value: string;
-  onChange: (value: string) => void;
-  onSubmit: () => void;
-  buttonLabel: string;
-  isBusy: boolean;
-  disabled?: boolean;
-}) {
-  return (
-    <View className="gap-3">
-      <TextInput
-        style={inputStyle}
-        value={value}
-        onChangeText={onChange}
-        placeholder="Enter team name"
-        placeholderTextColor={mflColors.textMuted}
-        maxLength={100}
-      />
-      <Button
-        variant="primary"
-        size="lg"
-        onPress={onSubmit}
-        isDisabled={isBusy || disabled}
-      >
-        {isBusy ? <Spinner size="sm" /> : <Button.Label>{buttonLabel}</Button.Label>}
-      </Button>
-    </View>
-  );
-}
-
-function SearchInput({
-  value,
-  onChange,
-  placeholder = 'Search members...',
-}: {
-  value: string;
-  onChange: (value: string) => void;
-  placeholder?: string;
-}) {
-  return (
-    <TextInput
-      style={inputStyle}
-      value={value}
-      onChangeText={onChange}
-      placeholder={placeholder}
-      placeholderTextColor={mflColors.textMuted}
-      autoCapitalize="none"
-    />
-  );
-}
-
-function UnallocatedPanel({
-  data,
-  targetTeamId,
-  selectedMemberIds,
-  search,
-  isBusy,
-  onClose,
-  onSearch,
-  onTargetTeam,
-  onToggleMember,
-  onAddSelected,
-}: {
-  data: TeamManagementData;
-  targetTeamId: string;
-  selectedMemberIds: string[];
-  search: string;
-  isBusy: boolean;
-  onClose: () => void;
-  onSearch: (value: string) => void;
-  onTargetTeam: (teamId: string) => void;
-  onToggleMember: (memberId: string) => void;
-  onAddSelected: () => void;
-}) {
-  const filtered = searchMembers(data.members.unallocated, search);
-  return (
-    <TeamManagementPanel
-      title="Unallocated Members"
-      subtitle={`${data.members.unallocated.length} member(s) are not assigned to a team.`}
-      onClose={onClose}
-    >
-      <TeamChips
-        teams={data.teams}
-        selectedTeamId={targetTeamId}
-        onSelect={onTargetTeam}
-      />
-      <SearchInput value={search} onChange={onSearch} />
-      <SelectableMembers
-        members={filtered}
-        selectedIds={selectedMemberIds}
-        onToggle={onToggleMember}
-      />
-      <Button
-        variant="primary"
-        size="lg"
-        onPress={onAddSelected}
-        isDisabled={isBusy || !targetTeamId || selectedMemberIds.length === 0}
-      >
-        {isBusy ? <Spinner size="sm" /> : <Button.Label>Add Members</Button.Label>}
-      </Button>
-    </TeamManagementPanel>
-  );
-}
-
-function AddMembersPanel({
-  team,
-  unallocatedMembers,
-  leagueCapacity,
-  maxTeams,
-  selectedMemberIds,
-  search,
-  isBusy,
-  onClose,
-  onSearch,
-  onToggleMember,
-  onAddSelected,
-}: {
-  team: ManagedTeam;
-  unallocatedMembers: ManagedLeagueMember[];
-  leagueCapacity: number;
-  maxTeams: number;
-  selectedMemberIds: string[];
-  search: string;
-  isBusy: boolean;
-  onClose: () => void;
-  onSearch: (value: string) => void;
-  onToggleMember: (memberId: string) => void;
-  onAddSelected: () => void;
-}) {
-  const capacity = getPerTeamCapacity({ leagueCapacity, maxTeams });
-  const remaining = Math.max(0, capacity - team.member_count);
-  const filtered = searchMembers(unallocatedMembers, search);
-
-  return (
-    <TeamManagementPanel
-      title={`Add Members to ${team.team_name}`}
-      subtitle={`${remaining} slot(s) remaining. ${unallocatedMembers.length} unallocated member(s).`}
-      onClose={onClose}
-    >
-      <SearchInput value={search} onChange={onSearch} />
-      <SelectableMembers
-        members={filtered}
-        selectedIds={selectedMemberIds}
-        onToggle={onToggleMember}
-      />
-      <Button
-        variant="primary"
-        size="lg"
-        onPress={onAddSelected}
-        isDisabled={isBusy || selectedMemberIds.length === 0 || remaining <= 0}
-      >
-        {isBusy ? <Spinner size="sm" /> : <Button.Label>Add Members</Button.Label>}
-      </Button>
-    </TeamManagementPanel>
-  );
-}
-
-function SelectableMembers({
-  members,
-  selectedIds,
-  onToggle,
-}: {
-  members: ManagedLeagueMember[];
-  selectedIds: string[];
-  onToggle: (memberId: string) => void;
-}) {
-  if (members.length === 0) {
-    return (
-      <Card className="p-4">
-        <AppText className="text-sm text-muted text-center">No members found.</AppText>
-      </Card>
-    );
-  }
-  return (
-    <View className="gap-2">
-      {members.map((member) => (
-        <MemberListRow
-          key={member.league_member_id}
-          name={member.username}
-          email={member.email}
-          roles={member.roles}
-          points={member.points}
-          selected={selectedIds.includes(member.league_member_id)}
-          onPress={() => onToggle(member.league_member_id)}
-        />
-      ))}
-    </View>
-  );
-}
-
-function TeamChips({
-  teams,
-  selectedTeamId,
-  onSelect,
-}: {
-  teams: ManagedTeam[];
-  selectedTeamId: string;
-  onSelect: (teamId: string) => void;
-}) {
-  return (
-    <View className="flex-row flex-wrap gap-2">
-      {teams.map((team) => (
-        <ActionButton
-          key={team.team_id}
-          label={team.team_name}
-          icon={selectedTeamId === team.team_id ? 'check' : 'circle'}
-          variant={selectedTeamId === team.team_id ? 'primary' : 'secondary'}
-          onPress={() => onSelect(team.team_id)}
-        />
-      ))}
-    </View>
-  );
-}
-
-function MembersPanel({
-  team,
-  teams,
-  members,
-  isHost,
-  isLoading,
-  search,
-  memberToMove,
-  targetTeamId,
-  isBusy,
-  onClose,
-  onSearch,
-  onMoveMember,
-  onTargetTeam,
-  onConfirmMove,
-  onUnassign,
-  onRemoveFromLeague,
-}: {
-  team: ManagedTeam;
-  teams: ManagedTeam[];
-  members: ManagedTeamMember[];
-  isHost: boolean;
-  isLoading: boolean;
-  search: string;
-  memberToMove: ManagedTeamMember | null;
-  targetTeamId: string;
-  isBusy: boolean;
-  onClose: () => void;
-  onSearch: (value: string) => void;
-  onMoveMember: (member: ManagedTeamMember | null) => void;
-  onTargetTeam: (teamId: string) => void;
-  onConfirmMove: () => void;
-  onUnassign: (member: ManagedTeamMember) => void;
-  onRemoveFromLeague: (member: ManagedTeamMember) => void;
-}) {
-  const filtered = searchMembers(members, search);
-  return (
-    <TeamManagementPanel
-      title={`${team.team_name} - Members`}
-      subtitle={`${members.length} member(s) in this team${
-        team.captain ? ` | Captain: ${team.captain.username}` : ''
-      }`}
-      onClose={onClose}
-    >
-      <SearchInput value={search} onChange={onSearch} />
-      {isLoading ? (
-        <View className="items-center py-6 gap-2">
-          <Spinner size="sm" />
-          <AppText className="text-sm text-muted">Loading members...</AppText>
-        </View>
-      ) : filtered.length === 0 ? (
-        <Card className="p-4">
-          <AppText className="text-sm text-muted text-center">No members found.</AppText>
-        </Card>
-      ) : (
-        <View className="gap-2">
-          {filtered.map((member) => (
-            <View key={member.league_member_id} className="gap-2">
-              <MemberListRow
-                name={member.username}
-                email={member.email}
-                roles={member.roles}
-                captain={member.is_captain}
-                points={member.points}
-              />
-              {isHost ? (
-                <View className="flex-row flex-wrap gap-2">
-                  <ActionButton
-                    label="Move"
-                    icon="arrow-right"
-                    onPress={() => onMoveMember(member)}
-                  />
-                  <ActionButton
-                    label="Unassign"
-                    icon="log-out"
-                    onPress={() => onUnassign(member)}
-                  />
-                  <ActionButton
-                    label="Remove from League"
-                    icon="trash-2"
-                    variant="danger"
-                    onPress={() => onRemoveFromLeague(member)}
-                  />
-                </View>
-              ) : null}
-            </View>
-          ))}
-        </View>
-      )}
-      {memberToMove ? (
-        <Card className="p-3 gap-3">
-          <AppText className="text-sm font-semibold text-foreground">
-            Move {memberToMove.username}
-          </AppText>
-          <TeamChips
-            teams={teams.filter((otherTeam) => otherTeam.team_id !== team.team_id)}
-            selectedTeamId={targetTeamId}
-            onSelect={onTargetTeam}
-          />
-          <Button
-            variant="primary"
-            size="md"
-            onPress={onConfirmMove}
-            isDisabled={isBusy || !targetTeamId}
-          >
-            {isBusy ? <Spinner size="sm" /> : <Button.Label>Move Member</Button.Label>}
-          </Button>
-        </Card>
-      ) : null}
-    </TeamManagementPanel>
-  );
-}
-
-function CaptainPanel({
-  team,
-  members,
-  isLoading,
-  selectedMemberIds,
-  search,
-  isBusy,
-  onClose,
-  onSearch,
-  onSelect,
-  onAssign,
-  onRemove,
-}: {
-  team: ManagedTeam;
-  members: ManagedTeamMember[];
-  isLoading: boolean;
-  selectedMemberIds: string[];
-  search: string;
-  isBusy: boolean;
-  onClose: () => void;
-  onSearch: (value: string) => void;
-  onSelect: (userId: string) => void;
-  onAssign: () => void;
-  onRemove: () => void;
-}) {
-  const filtered = searchMembers(members, search);
-  const currentCaptain = members.find((member) => member.is_captain);
-  return (
-    <TeamManagementPanel
-      title={`Assign Captain for ${team.team_name}`}
-      subtitle="The captain can validate submissions from team members."
-      onClose={onClose}
-    >
-      <SearchInput value={search} onChange={onSearch} placeholder="Search team members..." />
-      {isLoading ? (
-        <Spinner size="sm" />
-      ) : (
-        <View className="gap-2">
-          {filtered.map((member) => (
-            <MemberListRow
-              key={member.user_id}
-              name={member.username}
-              email={member.email}
-              roles={member.roles}
-              captain={member.is_captain}
-              selected={selectedMemberIds[0] === member.user_id}
-              onPress={() => onSelect(member.user_id)}
-            />
-          ))}
-        </View>
-      )}
-      <View className="flex-row gap-3">
-        {currentCaptain ? (
-          <Button variant="secondary" size="lg" onPress={onRemove} isDisabled={isBusy} className="flex-1">
-            <Button.Label>Remove Captain</Button.Label>
-          </Button>
-        ) : null}
-        <Button
-          variant="primary"
-          size="lg"
-          onPress={onAssign}
-          isDisabled={isBusy || !selectedMemberIds[0]}
-          className="flex-1"
-        >
-          {isBusy ? <Spinner size="sm" /> : <Button.Label>Assign Captain</Button.Label>}
-        </Button>
-      </View>
-    </TeamManagementPanel>
-  );
-}
-
-function ViceCaptainPanel({
-  team,
-  members,
-  isLoading,
-  search,
-  isBusy,
-  onClose,
-  onSearch,
-  onAssign,
-  onRemove,
-}: {
-  team: ManagedTeam;
-  members: ManagedTeamMember[];
-  isLoading: boolean;
-  search: string;
-  isBusy: boolean;
-  onClose: () => void;
-  onSearch: (value: string) => void;
-  onAssign: (userId: string) => void;
-  onRemove: (userId: string) => void;
-}) {
-  const current = members.filter((member) => member.roles.includes('vice_captain'));
-  const eligible = searchMembers(
-    members.filter(
-      (member) =>
-        !member.roles.includes('captain') && !member.roles.includes('vice_captain'),
-    ),
-    search,
-  );
-
-  return (
-    <TeamManagementPanel
-      title={`Vice Captains - ${team.team_name}`}
-      subtitle="Vice captains can validate submissions, manage team members, and act for the captain."
-      onClose={onClose}
-    >
-      {current.length > 0 ? (
-        <View className="gap-2">
-          <SectionLabel label="Current Vice Captains" />
-          {current.map((member) => (
-            <MemberListRow
-              key={member.user_id}
-              name={member.username}
-              email={member.email}
-              roles={member.roles}
-              right={
-                <ActionButton
-                  label="Remove"
-                  icon="x"
-                  variant="danger"
-                  disabled={isBusy}
-                  onPress={() => onRemove(member.user_id)}
-                />
-              }
-            />
-          ))}
-        </View>
-      ) : null}
-      <SearchInput value={search} onChange={onSearch} placeholder="Search members to add..." />
-      {isLoading ? (
-        <Spinner size="sm" />
-      ) : eligible.length === 0 ? (
-        <Card className="p-4">
-          <AppText className="text-sm text-muted text-center">No eligible members.</AppText>
-        </Card>
-      ) : (
-        <View className="gap-2">
-          {eligible.map((member) => (
-            <MemberListRow
-              key={member.user_id}
-              name={member.username}
-              email={member.email}
-              roles={member.roles}
-              right={
-                <ActionButton
-                  label="Add"
-                  icon="plus"
-                  disabled={isBusy}
-                  onPress={() => onAssign(member.user_id)}
-                />
-              }
-            />
-          ))}
-        </View>
-      )}
-    </TeamManagementPanel>
-  );
-}
-
-function GovernorsPanel({
-  members,
-  hostUserId,
-  currentGovernorIds,
-  selectedGovernorIds,
-  search,
-  isBusy,
-  onClose,
-  onSearch,
-  onToggle,
-  onSave,
-}: {
-  members: ManagedLeagueMember[];
-  hostUserId: string;
-  currentGovernorIds: string[];
-  selectedGovernorIds: string[];
-  search: string;
-  isBusy: boolean;
-  onClose: () => void;
-  onSearch: (value: string) => void;
-  onToggle: (userId: string) => void;
-  onSave: () => void;
-}) {
-  const eligible = searchMembers(
-    members.filter((member) => member.user_id !== hostUserId),
-    search,
-  );
-  return (
-    <TeamManagementPanel
-      title="Manage Governors"
-      subtitle="Governors have oversight of all teams and can validate any submission."
-      onClose={onClose}
-    >
-      <SearchInput value={search} onChange={onSearch} placeholder="Search league members..." />
-      <View className="gap-2">
-        {eligible.map((member) => (
-          <MemberListRow
-            key={member.user_id}
-            name={member.username}
-            email={member.email}
-            roles={member.roles}
-            points={member.points}
-            selected={selectedGovernorIds.includes(member.user_id)}
-            onPress={() => onToggle(member.user_id)}
-          />
-        ))}
-      </View>
-      <AppText className="text-xs text-muted">
-        Current governors selected: {currentGovernorIds.length}
-      </AppText>
-      <Button variant="primary" size="lg" onPress={onSave} isDisabled={isBusy}>
-        {isBusy ? <Spinner size="sm" /> : <Button.Label>Save Changes</Button.Label>}
-      </Button>
-    </TeamManagementPanel>
-  );
-}
-
-function InvitePanel({
-  leagueName,
-  team,
-  inviteCode,
-  memberCount,
-  maxCapacity,
-  onClose,
-}: {
-  leagueName: string;
-  team: ManagedTeam | null;
-  inviteCode: string;
-  memberCount: number;
-  maxCapacity: number;
-  onClose: () => void;
-}) {
-  const invitePath = team ? `/invite/team/${inviteCode}` : `/invite/${inviteCode}`;
-  const label = team ? `Join ${team.team_name}` : `Join ${leagueName}`;
-  const shareText = team
-    ? `Join my team "${team.team_name}" in ${leagueName} on MyFitnessLeague. Use code: ${inviteCode}`
-    : `Join ${leagueName} on MyFitnessLeague. Use code: ${inviteCode}`;
-
-  return (
-    <TeamManagementPanel
-      title={team ? 'Invite to Team' : 'League Invite'}
-      subtitle={`${memberCount} / ${maxCapacity} members`}
-      onClose={onClose}
-    >
-      <Card className="p-4 gap-2">
-        <AppText className="text-xs text-muted">Invite Code</AppText>
-        <AppText className="text-2xl font-bold text-foreground text-center tracking-widest">
-          {inviteCode || 'No code'}
-        </AppText>
-      </Card>
-      <AppText className="text-xs text-muted">Path: {invitePath}</AppText>
-      <Button
-        variant="primary"
-        size="lg"
-        onPress={() =>
-          Share.share({
-            title: label,
-            message: `${shareText}\n${invitePath}`,
-          })
-        }
-        isDisabled={!inviteCode}
-      >
-        <Button.Label>Share Invite</Button.Label>
-      </Button>
-    </TeamManagementPanel>
-  );
 }
